Update : value, gas_used field type altered: (IntegerField => PositiveBigIntegerField)

+++ scanner_daemon/scanner_backend/models/transaction.py
@@ -22,8 +22,8 @@ class Transaction(models.Model):
     related_sender = models.ForeignKey(DerivedWallet, blank=True, null=True, on_delete=models.CASCADE, related_name='sender')
     related_recipient = models.ForeignKey(DerivedWallet, blank=True, null=True, on_delete=models.CASCADE, related_name='recipient')
 
-    value = models.IntegerField(default=0)  # 18 Decimals applied(BNB)
-    gas_used = models.IntegerField(default=0)
+    value = models.PositiveBigIntegerField(default=0)  # 18 Decimals applied(BNB)
+    gas_used = models.PositiveBigIntegerField(default=0)
 
     sender_address = models.CharField(max_length=42, validators=[MinLengthValidator(42)], blank=True, null=True)  # Hash Address
     recipient_address = models.CharField(max_length=42, validators=[MinLengthValidator(42)], blank=True, null=True)
